+++ react/typings/vtex.totalizer-translator.d.ts
@@ -1,0 +1,2 @@
+declare module 'vtex.totalizer-translator'
+declare module 'vtex.totalizer-translator/TranslateTotalizer'